Update README.md
A personal portfolio is a collection of works, projects, achievements, and experiences that showcase an individual's skills, talents, and accomplishments. It serves as a representation of someone's professional or creative journey and is often used to demonstrate their expertise to potential employers, clients, or collaborators. Personal portfolios can take various forms, such as a physical portfolio, a website, a digital document, or a multimedia presentation. Typically, a personal portfolio includes examples of past work, a resume or curriculum vitae, testimonials, and sometimes a personal statement or bio to provide context and background information. The content of a personal portfolio can vary widely depending on the individual's field, interests, and goals.